+++ admin/drivers/mssql.inc.php
@@ -872,7 +872,7 @@ WHERE sys1.xtype = 'TR' AND sys2.name = " . q($table)
 		return "TRUNCATE TABLE " . table($table);
 	}
 
-	function use_sql($database) {
+	function use_sql($database, $style = "") {
 		return "USE " . idf_escape($database);
 	}